move binary sensor related constants
