+++ homeassistant/components/synology_dsm/const.py
@@ -3,17 +3,10 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 
-from synology_dsm.api.core.security import SynoCoreSecurity
-from synology_dsm.api.core.upgrade import SynoCoreUpgrade
-from synology_dsm.api.storage.storage import SynoStorage
 from synology_dsm.api.surveillance_station.const import SNAPSHOT_PROFILE_BALANCED
 
-from homeassistant.components.binary_sensor import (
-    BinarySensorDeviceClass,
-    BinarySensorEntityDescription,
-)
 from homeassistant.const import Platform
-from homeassistant.helpers.entity import EntityCategory, EntityDescription
+from homeassistant.helpers.entity import EntityDescription
 
 DOMAIN = "synology_dsm"
 PLATFORMS = [
@@ -71,50 +64,3 @@ class SynologyDSMRequiredKeysMixin:
 @dataclass
 class SynologyDSMEntityDescription(EntityDescription, SynologyDSMRequiredKeysMixin):
     """Generic Synology DSM entity description."""
-
-
-@dataclass
-class SynologyDSMBinarySensorEntityDescription(
-    BinarySensorEntityDescription, SynologyDSMEntityDescription
-):
-    """Describes Synology DSM binary sensor entity."""
-
-
-# Binary sensors
-UPGRADE_BINARY_SENSORS: tuple[SynologyDSMBinarySensorEntityDescription, ...] = (
-    SynologyDSMBinarySensorEntityDescription(
-        # Deprecated, scheduled to be removed in 2022.6 (#68664)
-        api_key=SynoCoreUpgrade.API_KEY,
-        key="update_available",
-        name="Update Available",
-        entity_registry_enabled_default=False,
-        device_class=BinarySensorDeviceClass.UPDATE,
-        entity_category=EntityCategory.DIAGNOSTIC,
-    ),
-)
-
-SECURITY_BINARY_SENSORS: tuple[SynologyDSMBinarySensorEntityDescription, ...] = (
-    SynologyDSMBinarySensorEntityDescription(
-        api_key=SynoCoreSecurity.API_KEY,
-        key="status",
-        name="Security Status",
-        device_class=BinarySensorDeviceClass.SAFETY,
-    ),
-)
-
-STORAGE_DISK_BINARY_SENSORS: tuple[SynologyDSMBinarySensorEntityDescription, ...] = (
-    SynologyDSMBinarySensorEntityDescription(
-        api_key=SynoStorage.API_KEY,
-        key="disk_exceed_bad_sector_thr",
-        name="Exceeded Max Bad Sectors",
-        device_class=BinarySensorDeviceClass.SAFETY,
-        entity_category=EntityCategory.DIAGNOSTIC,
-    ),
-    SynologyDSMBinarySensorEntityDescription(
-        api_key=SynoStorage.API_KEY,
-        key="disk_below_remain_life_thr",
-        name="Below Min Remaining Life",
-        device_class=BinarySensorDeviceClass.SAFETY,
-        entity_category=EntityCategory.DIAGNOSTIC,
-    ),
-)
